% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_txt_tables.R
\name{get_txt_page}
\alias{get_txt_page}
\title{Get lines of table txt_page such that the column 'variable' contains the text pattern 'value'}
\usage{
get_txt_page(conn, value, variable = "text_en")
}
\arguments{
\item{conn}{the connection to the GloUrb package}

\item{value}{the text pattern that is searched in the lines of the txt_page table.}

\item{variable}{the variable in which to look for the text pattern. Defaults to text_en.}
}
\value{
the filtered txt_page table
}
\description{
Get lines of table txt_page such that the column 'variable' contains the text pattern 'value'
}
\examples{
conn=glourbi::connect_to_glourb()
destroy_in_text_en=get_txt_page(conn,value="destroy")
dam_in_title=get_txt_page(conn,variable="title",value="dam")
damordyke_in_title=get_txt_page(conn,variable="title",value="dam|dyke")
}
