% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_txt_tables.R
\name{get_txt_segment}
\alias{get_txt_segment}
\title{Get lines of table txt_segment such that the column 'variable' contains the text pattern 'value'}
\usage{
get_txt_segment(conn, value, variable = "text")
}
\arguments{
\item{conn}{the connection to the GloUrb package}

\item{value}{the text pattern that is searched in the lines of the txt_page table.}

\item{variable}{the variable in which to look for the text pattern. Defaults to text.}
}
\value{
the filtered txt_segment table
}
\description{
Get lines of table txt_segment such that the column 'variable' contains the text pattern 'value'
}
\examples{
conn=glourbi::connect_to_glourb()
get_txt_segment(conn,value)
}
